# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Entitlement {
  premium: boolean;
  /** @nullable */
  tier?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  grantedAt?: string | null;
}

export interface Habit {
  id: string;
  name: string;
  color: string;
  position: number;
}

export interface HabitInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  color: string;
}

export interface HabitUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  color?: string;
  position?: number;
}

export interface HabitReorder {
  ids: string[];
}

export interface Completion {
  habitId: string;
  date: string;
}

export interface CompletionToggle {
  habitId: string;
  date: string;
}

export interface ToggleResult {
  habitId: string;
  date: string;
  completed: boolean;
}

export interface HabitImport {
  id: string;
  name: string;
  color: string;
  position: number;
}

export interface SyncImport {
  habits: HabitImport[];
  completions: Completion[];
}

export interface SyncState {
  habits: Habit[];
  completions: Completion[];
}