# lib/api-zod/src/generated/types/entitlement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Entitlement {
  premium: boolean;
  /** @nullable */
  tier?: string | null;
  /** @nullable */
  source?: string | null;
  /** @nullable */
  grantedAt?: string | null;
}