# lib/api-zod/src/generated/types/habitInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HabitInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  color: string;
}