# lib/api-zod/src/generated/types/syncImport.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Completion } from './completion';
import type { HabitImport } from './habitImport';

export interface SyncImport {
  habits: HabitImport[];
  completions: Completion[];
}