# lib/api-zod/src/generated/types/syncState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Completion } from './completion';
import type { Habit } from './habit';

export interface SyncState {
  habits: Habit[];
  completions: Completion[];
}